refator : 느슨한 결합 , Service & Impl
