+++ src/main/java/com/book_everywhere/book/controller/BookController.java
@@ -32,7 +32,7 @@ public class BookController {
     //유저 책 조회
     @GetMapping("/api/mypage/{socialId}")
     public CMRespDto<?> findOneBookWithUser(@PathVariable Long socialId) {
-        List<BookDto> result = bookService.findAllBookOneUser(socialId);
+        List<BookDto> result = bookService.유저책조회(socialId);
         return new CMRespDto<>(HttpStatus.OK, result, "유저의 책 조회 완료");
     }
 }
