+++ src/main/resources/application.yml
@@ -14,6 +14,7 @@ spring:
   config:
     activate:
       on-profile: local
+
   security:
     oauth2:
       client:
